Add BaseModel
- 각 모델이 baseModel 을 상속받아 사용하도록 수정
